backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.26.2
thefuzz==0.19.0
python-Levenshtein==0.21.1
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
httpx==0.25.2
openai==1.3.7
python-dotenv==1.0.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

